requirements.txt
asgiref==3.3.1
defusedxml==0.6.0
diff-match-patch==20200713
Django==3.1.4
django-appconf==1.0.4
django-imagekit==4.0.2
django-import-export==2.5.0
django-widget-tweaks==1.4.8
et-xmlfile==1.0.1
gunicorn==20.0.4
jdcal==1.4.1
MarkupPy==1.14
odfpy==1.4.1
openpyxl==3.0.5
pilkit==2.0
Pillow==8.0.1
pytz==2020.5
PyYAML==5.3.1
six==1.15.0
sqlparse==0.4.1
tablib==3.0.0
xlrd==2.0.1
xlwt==1.3.0
